Now the dry_run options sets the new_tag param
The dry_run option did not set the new_tag output parameter and we could
not use the action to compute future tag when opening a PR due to the
pullRequest.merged check.

So now we can use the dry_run=true on the PR open to compute the tag
that would be used in the merge event.