Added milestone as optional input parameter
When using this action in a workflow together with an action that would automatically
attached a milestone to the PR, the `pull_request.milestone` might be null
or still to an old value. Therefore in this case we can use the new
input parameter `milestone` to overwrite the value from
pull_request.milestone.title.